moved frequent element function into new module
