+++ hw4/moduleSort/frequentElement.h
@@ -1,0 +1,9 @@
+#pragma once
+
+#include <stdbool.h>
+
+// This function checks if programm finds frequent element correctly
+void frequentElementCheck(int* inputArray, int arraySize, int* maximumCounter, int* frequentElement);
+
+// This function checks if funstion finds frequent element correctly
+bool testFrequentCase();
