+++ hw4/moduleSort/qsortTests.c
@@ -22,33 +22,6 @@ bool qsortCheck(int* sortingArray, int* correctArray, int arraySize)
     return true;
 }
 
-void frequentElementCheck(int* inputArray, int arraySize, int* maximumCounter, int* frequentElement)
-{
-    int currentCounter = 1;
-    *maximumCounter = 0;
-    for (int i = 1; i < arraySize; ++i)
-    {
-        if (inputArray[i - 1] != inputArray[i])
-        {
-            if (currentCounter > *maximumCounter)
-            {
-                *maximumCounter = currentCounter;
-                currentCounter = 1;
-                *frequentElement = inputArray[i - 1];
-            }
-        }
-        else
-        {
-            ++currentCounter;
-        }
-    }
-    if (currentCounter > *maximumCounter)
-    {
-        *maximumCounter = currentCounter;
-        *frequentElement = inputArray[arraySize - 1];
-    }
-}
-
 bool testCorrectCase()
 {
     int sortingArray[ARRAY_SIZE] = { -4, 3, 44, -321, -4, 40, 41, -42, -35, 6, 60, 79, -53, 5, 5, -5, 17, 111, 0, -9 };
@@ -60,9 +33,5 @@ bool testCorrectCase()
     {
         return false;
     }
-
-    int maximumCounter = 1;
-    int frequentElement = sortingArray[0];
-    frequentElementCheck(sortingArray, ARRAY_SIZE, &maximumCounter, &frequentElement);
-    return frequentElement == 5 || frequentElement == -4 || maximumCounter == 2;
+    return true;
 }
